Finish moving to base protocol class. Rename base client and tcp client. Refs #72.
